Add namespace convert api
